Add test for Commit Resolver
