+++ packages/amplication-server/src/core/app/app.resolver.spec.ts
@@ -27,6 +27,8 @@ import {
   EnumPendingChangeResourceType
 } from 'amplication-data/dist/models';
 import { mockGqlAuthGuardCanActivate } from '../../../test/gql-auth-mock';
+import { CommitResolver } from './commit.resolver';
+import { UserService } from '../user/user.service';
 
 const EXAMPLE_APP_ID = 'exampleAppId';
 const EXAMPLE_NAME = 'exampleName';
@@ -352,6 +354,18 @@ const GET_COMMITS_QUERY = gql`
   }
 `;
 
+const USER_MUTATION = gql`
+  mutation($message: String!, $appId: String!) {
+    commit(data: { message: $message, app: { connect: { id: $appId } } }) {
+      user {
+        id
+        createdAt
+        updatedAt
+      }
+    }
+  }
+`;
+
 const appMock = jest.fn(() => {
   return EXAMPLE_APP;
 });
@@ -385,6 +399,7 @@ const findManyBuildMock = jest.fn(() => {
 const findManyEnvironmentsMock = jest.fn(() => {
   return [EXAMPLE_ENVIRONMENT];
 });
+const userServiceFindUserMock = jest.fn(() => EXAMPLE_USER);
 
 const mockCanActivate = jest.fn(mockGqlAuthGuardCanActivate(EXAMPLE_USER));
 
@@ -397,6 +412,7 @@ describe('AppResolver', () => {
     const moduleFixture: TestingModule = await Test.createTestingModule({
       providers: [
         AppResolver,
+        CommitResolver,
         {
           provide: AppService,
           useClass: jest.fn(() => ({
@@ -408,6 +424,12 @@ describe('AppResolver', () => {
             discardPendingChanges: discardPendingChangesMock,
             getPendingChanges: getPendingChangesMock,
             getCommits: getCommitsMock
+          }))
+        },
+        {
+          provide: UserService,
+          useClass: jest.fn(() => ({
+            findUser: userServiceFindUserMock
           }))
         },
         {
@@ -774,5 +796,30 @@ describe('AppResolver', () => {
     });
     expect(getCommitsMock).toBeCalledTimes(1);
     expect(getCommitsMock).toBeCalledWith({});
+  });
+
+  /* Move to independent Commit Resolver Spec if Commit Resolver ever becomes independent */
+  it('should find committing user', async () => {
+    const res = await apolloClient.query({
+      query: USER_MUTATION,
+      variables: {
+        message: EXAMPLE_MESSAGE,
+        appId: EXAMPLE_APP_ID
+      }
+    });
+    expect(res.errors).toBeUndefined();
+    expect(res.data).toEqual({
+      commit: {
+        user: {
+          ...EXAMPLE_USER,
+          createdAt: EXAMPLE_USER.createdAt.toISOString(),
+          updatedAt: EXAMPLE_USER.updatedAt.toISOString()
+        }
+      }
+    });
+    expect(userServiceFindUserMock).toBeCalledTimes(1);
+    expect(userServiceFindUserMock).toBeCalledWith({
+      where: { id: EXAMPLE_USER_ID }
+    });
   });
 });
